We have a SOCKET PROBLEM (see description)
it seemed to work past days, but now, when i try to use socket.io on a static enviroment (like javascripts/index.js is), it does work no more.

The teacher always used VIEWS in the socket examples, so the problem persists and I found out no solution for hours yet.